// case_4_on_transaction/sentio/src/schema/schema.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type { String, Int, BigInt, Float, ID, Bytes, Timestamp, Boolean, Int8 } from '@sentio/sdk/store'
import { Entity, Required, One, Many, Column, ListColumn, AbstractEntity, getStore, UpdateValues } from '@sentio/sdk/store'
import { BigDecimal } from '@sentio/bigdecimal'
import { DatabaseSchema } from '@sentio/sdk'







interface GasSpentConstructorInput {
  id: ID;
  from: String;
  to: String;
  gasValue: BigInt;
  gasUsed: BigInt;
  gasPrice: BigInt;
  effectiveGasPrice?: BigInt;
  blockNumber: BigInt;
  transactionHash?: Bytes;
  indexedAt: BigInt;
}
@Entity("GasSpent")
export class GasSpent extends AbstractEntity  {

	@Required
	@Column("ID")
	id: ID

	@Required
	@Column("String")
	from: String

	@Required
	@Column("String")
	to: String

	@Required
	@Column("BigInt")
	gasValue: BigInt

	@Required
	@Column("BigInt")
	gasUsed: BigInt

	@Required
	@Column("BigInt")
	gasPrice: BigInt

	@Column("BigInt")
	effectiveGasPrice?: BigInt

	@Required
	@Column("BigInt")
	blockNumber: BigInt

	@Column("Bytes")
	transactionHash?: Bytes

	@Required
	@Column("BigInt")
	indexedAt: BigInt
  constructor(data: GasSpentConstructorInput) {super()}
  
  
  static update(values: UpdateValues<GasSpentConstructorInput>): Promise<void> {
    return getStore().update(GasSpent, values)
  }
}


const source = `type GasSpent @entity {
  id: ID!                   # Transaction hash
  from: String!             # Sender address 
  to: String!               # Recipient address
  gasValue: BigInt!         # Gas cost (gasUsed * gasPrice)
  gasUsed: BigInt!          # Gas used by the transaction
  gasPrice: BigInt!         # Base gas price
  effectiveGasPrice: BigInt # Effective gas price (for EIP-1559 transactions)
  blockNumber: BigInt!      # Block where transaction was executed
  transactionHash: Bytes    # Transaction hash as binary
  indexedAt: BigInt!
}

`
DatabaseSchema.register({
  source,
  entities: {
    "GasSpent": GasSpent
  }
})